Struct rubefunge_93::befunge::Program

pub struct Program {
    // some fields omitted
}

The Befunge program

Contains the user specified values and the instructions.

Methods

impl Program

fn new(values: Vec<u32>, instructions: Vec<Vec<char>>) -> Program

Create a new Befunge program with specified user values and instructions.

fn lines(&self) -> usize

Number of lines in the Befunge program.

fn chars_in_line(&self, line: usize) -> usize

Number of characters in the specified line.

fn next_value(&mut self) -> u32

Gets the next user values from the front of the Vec.

fn get_instruction_char(&self, pos: [usize; 2]) -> char

Gets the instruction character at the given position.

fn set_instruction_char(&mut self, pos: [usize; 2], c: char)

Sets the instruction character at the given position.